Revert "(feat) refactor auth features into authController.js"
